Move LoginNoWPcomAccountFoundFragment to the error package
